+++ core/src/main/kotlin/org/pastalab/fray/core/RunContext.kt
@@ -35,7 +35,6 @@ import org.pastalab.fray.core.concurrency.primitives.SignalContext
 import org.pastalab.fray.core.concurrency.primitives.StampedLockContext
 import org.pastalab.fray.core.concurrency.primitives.WriteLockContext
 import org.pastalab.fray.core.scheduler.FrayIdeaPluginScheduler
-import org.pastalab.fray.core.scheduler.SURWScheduler
 import org.pastalab.fray.core.syncurity.SyncurityEvaluationContext
 import org.pastalab.fray.core.syncurity.SyncurityEvaluationDelegate
 import org.pastalab.fray.core.utils.Utils.verifyOrReport
@@ -223,7 +222,7 @@ class RunContext(val config: Configuration) {
     mainThreadId = t.id
     registeredThreads[t.id] = ThreadContext(t, registeredThreads.size, this, -1)
     registeredThreads[t.id]?.state = ThreadState.Runnable
-    config.scheduleObservers.forEach { it.onExecutionStart() }
+    config.testStatusObservers.forEach { it.onExecutionStart() }
     scheduleNextOperation(true)
   }
 
@@ -236,7 +235,7 @@ class RunContext(val config: Configuration) {
     latchManager.done()
 
     registeredThreads.clear()
-    config.scheduleObservers.forEach { it.onExecutionDone(bugFound) }
+    config.testStatusObservers.forEach { it.onExecutionDone(bugFound) }
     hashCodeMapper.done(false)
     nanoTime = TimeUnit.SECONDS.toNanos(1577768400)
   }
@@ -1054,16 +1053,12 @@ class RunContext(val config: Configuration) {
     }
 
     val nextThread =
-        if (enabledOperations.size == 1 && config.scheduler !is SURWScheduler) {
+        try {
+          config.scheduler.scheduleNextOperation(
+              enabledOperations, registeredThreads.values.toList())
+        } catch (e: Throwable) {
+          reportError(e)
           enabledOperations.first()
-        } else {
-          try {
-            config.scheduler.scheduleNextOperation(
-                enabledOperations, registeredThreads.values.toList())
-          } catch (e: Throwable) {
-            reportError(e)
-            enabledOperations.first()
-          }
         }
     config.scheduleObservers.forEach {
       it.onNewSchedule(registeredThreads.values.toList().toThreadInfos(), nextThread.toThreadInfo())