+++ core/src/main/kotlin/org/pastalab/fray/core/scheduler/FrayIdeaPluginScheduler.kt
@@ -1,5 +1,6 @@
 package org.pastalab.fray.core.scheduler
 
+import kotlinx.serialization.Contextual
 import kotlinx.serialization.Serializable
 import org.pastalab.fray.core.ThreadContext
 import org.pastalab.fray.core.debugger.DebuggerRegistry
@@ -7,19 +8,26 @@ import org.pastalab.fray.core.debugger.DebuggerRegistry
 @Serializable
 class FrayIdeaPluginScheduler(val scheduler: Scheduler?) : Scheduler {
   @Transient val remoteScheduler = DebuggerRegistry.getRemoteScheduler()
+  @Contextual var previousScheduleDecision: ThreadContext? = null
 
   override fun scheduleNextOperation(
       threads: List<ThreadContext>,
       allThreads: List<ThreadContext>
   ): ThreadContext {
-    if (threads.size == 1) {
-      return threads.first()
-    }
     val thread = scheduler?.scheduleNextOperation(threads, allThreads)
-    val index =
-        remoteScheduler.scheduleNextOperation(
-            allThreads.map { it.toThreadInfo() }, thread?.toThreadInfo())
-    return allThreads[index]
+    return if (thread == null || thread.index != previousScheduleDecision?.index) {
+      previousScheduleDecision =
+          if (previousScheduleDecision == null || threads.size > 1 || scheduler != null) {
+            allThreads[
+                remoteScheduler.scheduleNextOperation(
+                    allThreads.map { it.toThreadInfo() }, thread?.toThreadInfo())]
+          } else {
+            threads.first()
+          }
+      previousScheduleDecision!!
+    } else {
+      thread
+    }
   }
 
   override fun nextIteration(): Scheduler {